feat(compat): support Unity 6000.5 and extend the matrix to 6000.6
Unity 6000.5 replaced 32-bit instance ids with a 64-bit EntityId whose upper
word is a per-session stamp. Truncating one to int produced ids that resolved
to nothing: object commands only kept working through the hierarchy-scan
fallback, and prefab/create, material texture references, and get-property on
a fresh id failed. Ids are now `long` on the wire through UnityObjectCompat;
6000.0-6000.4 keep their old numbers.

Also:
- The matrix and release validation request slots 6000.0-6000.6 (a missing
  slot falls back and is reported as covered). On 6000.5+ the matrix raises the
  navigation, collab-proxy, inputsystem, timeline, and visualscripting pins to
  the bundled versions and drops the removed com.unity.modules.vr module,
  which otherwise leaves an interactive editor on "Enter Safe Mode?".
- ensure_editor_running answers Unity startup dialogs on every poll; that
  loop is what actually waits on an editor parked on "Enter Safe Mode?" or
  "Packages with Errors", but the dialog policy only ran in wait_for_bridge.
  "Packages with Errors" is now recognised (Dismiss, never Dismiss Forever),
  exact button labels beat substring matches, and the handler traces what it
  saw under -v.
- The tests/run console guard skips log lines inside a test's
  LogAssert.ignoreFailingMessages window.
- One remaining direct GetInstanceID() call in the smoke tests uses the
  compat helper.

Validated: 123 Rust tests; bridge edit-mode 161/161 on 6000.4 and 6000.5 in
batch mode, and 161/161 through the live bridge on 6000.5.

+++ unity-package/com.ucp.bridge/Editor/Controllers/HierarchyController.cs
@@ -44,7 +44,7 @@ namespace UCP.Bridge
             // Optional parent
             if (p != null && p.TryGetValue("parent", out var parentObj))
             {
-                int parentId = Convert.ToInt32(parentObj);
+                long parentId = Convert.ToInt64(parentObj);
                 var parent = FindGameObject(parentId);
                 go.transform.SetParent(parent.transform, false);
             }
@@ -84,7 +84,7 @@ namespace UCP.Bridge
             if (p == null || !p.TryGetValue("instanceId", out var idObj))
                 throw new ArgumentException("Missing 'instanceId' parameter");
 
-            int instanceId = Convert.ToInt32(idObj);
+            long instanceId = Convert.ToInt64(idObj);
             var go = FindGameObject(instanceId);
             string name = go.name;
             var scene = go.scene;
@@ -107,14 +107,14 @@ namespace UCP.Bridge
             if (p == null || !p.TryGetValue("instanceId", out var idObj))
                 throw new ArgumentException("Missing 'instanceId' parameter");
 
-            int instanceId = Convert.ToInt32(idObj);
+            long instanceId = Convert.ToInt64(idObj);
             var go = FindGameObject(instanceId);
 
             Undo.SetTransformParent(go.transform, null, "UCP Reparent");
 
             if (p.TryGetValue("parent", out var parentObj) && parentObj != null)
             {
-                int parentId = Convert.ToInt32(parentObj);
+                long parentId = Convert.ToInt64(parentObj);
                 var parent = FindGameObject(parentId);
                 Undo.SetTransformParent(go.transform, parent.transform, "UCP Reparent");
             }
@@ -163,7 +163,7 @@ namespace UCP.Bridge
             // Instantiate from existing scene object (clone)
             else if (p.TryGetValue("sourceId", out var srcObj))
             {
-                int srcId = Convert.ToInt32(srcObj);
+                long srcId = Convert.ToInt64(srcObj);
                 source = FindGameObject(srcId);
             }
             else
@@ -183,7 +183,7 @@ namespace UCP.Bridge
             // Optional parent
             if (p.TryGetValue("parent", out var parentObj) && parentObj != null)
             {
-                int parentId = Convert.ToInt32(parentObj);
+                long parentId = Convert.ToInt64(parentObj);
                 var parent = FindGameObject(parentId);
                 instance.transform.SetParent(parent.transform, false);
             }
@@ -222,7 +222,7 @@ namespace UCP.Bridge
             if (!p.TryGetValue("type", out var typeObj) || typeObj == null)
                 throw new ArgumentException("Missing 'type' parameter");
 
-            int instanceId = Convert.ToInt32(idObj);
+            long instanceId = Convert.ToInt64(idObj);
             var go = FindGameObject(instanceId);
             string typeName = typeObj.ToString();
 
@@ -252,7 +252,7 @@ namespace UCP.Bridge
             if (!p.TryGetValue("type", out var typeObj) || typeObj == null)
                 throw new ArgumentException("Missing 'type' parameter");
 
-            int instanceId = Convert.ToInt32(idObj);
+            long instanceId = Convert.ToInt64(idObj);
             var go = FindGameObject(instanceId);
             string typeName = typeObj.ToString();
 
@@ -344,7 +344,7 @@ namespace UCP.Bridge
             return null;
         }
 
-        private static GameObject FindGameObject(int instanceId)
+        private static GameObject FindGameObject(long instanceId)
         {
             var obj = UnityObjectCompat.ResolveByInstanceId<GameObject>(instanceId);
             if (obj != null) return obj;
@@ -363,7 +363,7 @@ namespace UCP.Bridge
             throw new ArgumentException($"GameObject not found: {instanceId}");
         }
 
-        private static GameObject FindInHierarchy(GameObject go, int instanceId)
+        private static GameObject FindInHierarchy(GameObject go, long instanceId)
         {
             if (go.GetId() == instanceId) return go;
             for (int i = 0; i < go.transform.childCount; i++)
